refactor utils2 -> kodi_utils
- move getKodiVersion, and addMenuDirectoryItem to kodi_utils
